+++ changelog.md
@@ -12,6 +12,7 @@
 * Catch repeated commit exceptions properly.
 * Rejection of negative transaction fees. 
 * Handling of CORS with credentials. 
+* Handle case `Error.captureStackTrace` is undefined (Firefox).
 
 ### Changed